% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob-analytics.R
\docType{data}
\name{depth}
\alias{depth}
\title{Depth.}
\format{A data.frame consisting of the following fields:
\describe{
  \item{timestamp}{Time at which volume was added or removed.}
  \item{price}{Order book price level.}
  \item{volume}{Amount of remaining volume at this price level.}
  \item{side}{The side of the price level: \emph{bid} or \emph{ask}.}
}}
\description{
Price level depth (liquidity) through time.
}
\details{
The depth data.frame describes the amount of available volume for all price
levels in the limit order book through time. Each row corresponds to a limit
order event, in which volume has been added or removed.
}
\author{
phil
}
\seealso{
Other Limit.order.book.data: \code{\link{depth.summary}},
  \code{\link{events}}, \code{\link{trades}}
}
\keyword{datasets}

