% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob-analytics.R
\docType{data}
\name{depth.summary}
\alias{depth.summary}
\title{Depth summary.}
\format{A data.frame consisting of the following fields:
\describe{
  \item{timestamp}{Local timestamp corresponding to \code{\link{events}}.}
  \item{best.bid.price}{Best bid price.}
  \item{best.bid.vol}{Amount of volume available at the best bid.}
  \item{bid.vol25:500bps}{The amount of volume available for 20 25bps
percentiles below the best bid.}
  \item{best.ask.price}{The best ask price.}
  \item{best.ask.vol}{Amount of volume available at the best ask.}
  \item{ask.vol25:500bps}{The amount od volume available for 20 25bps
percentiles above the best ask.}
}}
\description{
Limit order book summary statistics.
}
\details{
Various summary statistics describing the state of the order book after
every limit order event. The metrics are intended to quantify the
\emph{shape} of the order book through time.
}
\seealso{
Other Limit order book data: \code{\link{depth}},
  \code{\link{events}}, \code{\link{trades}}
}
\author{
phil
}
\concept{Limit order book data}
\keyword{datasets}
