% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/depth.R
\name{filterDepth}
\alias{filterDepth}
\title{Filter priceLevelVolume (depth).}
\usage{
filterDepth(d, from, to)
}
\arguments{
\item{d}{Depth data (lob.data$depth).}

\item{from}{Beginning of range.}

\item{to}{End of range.}
}
\value{
Filtered depth data.
}
\description{
Given depth data calculated by priceLevelVolume(), filter so that depth
level changes are between a range: depth changes < or > range are discarded
and min,max timestamps remaining within the range are "clamped" - set to,
min,max of specified range.
}
\author{
phil
}

