% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/depth.R
\name{getSpread}
\alias{getSpread}
\title{Get the spread.}
\usage{
getSpread(depth.summary)
}
\arguments{
\item{depth.summary}{Depth summary data.}
}
\value{
Bid/Ask spread
}
\description{
Extracts the spread from the depth summary, removing any points in which a
change to bid/ask price/volume did not occur.
}
\examples{
with(lob.data, tail(getSpread(depth.summary)))
}
\author{
phil
}

