% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{loadData}
\alias{loadData}
\title{Load pre-processed data.}
\usage{
loadData(bin.file)
}
\arguments{
\item{bin.file}{File location.}
}
\value{
Limit order, trade and depth data structure.
}
\description{
Loads previously saved pre-processed data.
}
\author{
phil
}

