% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-processing.R
\name{loadEventData}
\alias{loadEventData}
\title{Read raw limit order event data from a CSV file.}
\usage{
loadEventData(file, price.digits = 2, volume.digits = 8)
}
\arguments{
\item{file}{Location of CSV file containing limit order events.}

\item{price.digits}{an integer indicating the number of decimal places in 'price' column of the CSV file}

\item{volume.digits}{an integer indicating the number of decimal places in 'volume' column of the CSV file}
}
\value{
A data.frame containing the raw limit order events data.
}
\description{
The function performs some data sanitisation: removing duplicate events
and ensuring that order events are in the appropriate order. The data are
ordered according to the life-cycle of a limit order:
    order id, volume (desc), action (created,changed,deleted)
}
\details{
In addition to returning the sanitised raw data, the function also
appends a column "fill.deltas" representing the change in volume for each
limit order event.
}
\author{
phil
}
\keyword{internal}
