% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microstructure2.R
\docType{data}
\name{lob.data}
\alias{lob.data}
\title{Example pre-processed data.}
\format{A list containing 4 data frames.}
\usage{
data(lob.data)
}
\description{
24 hours of Limit Order Book data obtained from the Bitstamp exchange on
2015-05-01.
}
\details{
\itemize{
  \item events. Limit order events.
  \item trades. Inferred trades (executions).
  \item depth.  Order book price level depth through time.
  \item depth.summary. Limit order book summary statistics.
}
}
\keyword{datasets}

