% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microstructure2.R
\docType{package}
\name{microstructure2}
\alias{microstructure2}
\alias{microstructure2-package}
\title{microstructure2.}
\description{
Limit order book event analysis.
}
\details{
Main functionality:
\itemize{
  \item Limit order book event processing and analysis.
  \item Visualise order book state and market impacts.
  \item Order book reconstruction.
}
}
\author{
phil nocturne \email{phil@parasec.net}
}
\references{
\url{http://parasec.net/transmission/order-book-visualisation}
}

