% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quotes-processing.R
\name{orderAggressiveness}
\alias{orderAggressiveness}
\title{Calculate order aggressiveness with respect to best bid or ask in BPS.}
\usage{
orderAggressiveness(events, depth.summary)
}
\arguments{
\item{events}{The events data.table.}

\item{depth.summary}{Order book summary statistics.}
}
\value{
The events data.table containing a new aggressiveness.bps column.
}
\description{
Added or deleted limit orders can be assigned a level of "aggressiveness"
with respect to the current best bid (ask) immediately before the order is
added or removed from the order book. Orders placed at the best bid (ask)
are assigned an aggressiveness of 0 BPS, a negative BPS below the best bid
(ask) and a positive BPS if placed inside the spread.
}
\author{
phil
}
\keyword{internal}

