% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order-book-reconstruction.R
\name{orderBook}
\alias{orderBook}
\title{Instantaneous limit order book reconstruction.}
\usage{
orderBook(events, tp = as.POSIXlt(Sys.time(), tz = "UTC"),
  max.price.levels = NULL, bps.range = 0, min.bid = 0, max.ask = Inf)
}
\arguments{
\item{events}{Limit order events data.frame.}

\item{tp}{Time point to re-construct order book at.}

\item{max.price.levels}{Max number of price levels to return.}

\item{bps.range}{Max depth to return +- BPS from best bid/ask.}

\item{min.bid}{Min bid to return.}

\item{max.ask}{Max ask to return.}
}
\value{
Limit Order Book structure:

\preformatted{
> tp <- as.POSIXct("2015-05-01 03:30:15.342", tz="UTC")
> orderBook(lob.data$events, tp, max.price.levels=5)
$timestamp
[1] "2015-05-01 03:30:15.342 UTC"

$asks
        id               timestamp  exchange.timestamp  price     volume  liquidity       bps
1 65613703 2015-05-01 03:29:53.127 2015-05-01 03:29:52 236.58  910229141 6341547077 2.1138968
2 65613655 2015-05-01 03:29:24.703 2015-05-01 03:29:24 236.56 1320000000 5431317936 1.2683381
3 65613700 2015-05-01 03:29:52.040 2015-05-01 03:29:51 236.55 1320000000 4111317936 0.8455587
4 65613698 2015-05-01 03:29:51.510 2015-05-01 03:29:51 236.54 1600000000 2791317936 0.4227794
5 65613712 2015-05-01 03:29:56.755 2015-05-01 03:29:56 236.53 1191317936 1191317936 0.0000000

$bids
        id               timestamp  exchange.timestamp  price    volume liquidity      bps
1 65613225 2015-05-01 03:27:44.450 2015-05-01 03:24:00 236.36  16154172  16154172 0.000000
2 65613681 2015-05-01 03:29:42.259 2015-05-01 03:29:41 236.31 200000000 216154172 2.115417
3 65613220 2015-05-01 03:23:57.963 2015-05-01 03:23:57 236.30 100000000 316154172 2.538501
4 65612978 2015-05-01 03:21:10.454 2015-05-01 03:21:10 236.28 100000000 416154172 3.384667
5 65612388 2015-05-01 03:13:08.808 2015-05-01 03:13:08 236.17 100000000 516154172 8.038585

    where        id = Limit order id.
          timestamp = Last update time (last time order was partially hit)
 exchange.timestamp = Order creation time.
              price = Limit price.
             volume = Remaining order volume.
          liquidity = Cumulative sum of volume up util this order (depth).
                bps = Distance in BPS from current best bid/ask.
}
}
\description{
Reconstructs a limit order book for a specific point in time.
}
\examples{
tp <- as.POSIXct("2015-05-01 04:25:15.342", tz="UTC")
orderBook(lob.data$events, max.price.levels=5)
}
\author{
phil
}

