% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.time.series}
\alias{plot.time.series}
\title{General time series plot.}
\usage{
\method{plot}{time.series}(timestamp, series, start.time = min(timestamp),
  end.time = max(timestamp), title = "time series", y.label = "series")
}
\arguments{
\item{timestamp}{POSIXct timestamps.}

\item{series}{The time series.}

\item{start.time,}{end.time, Inclusive start and end time of plot.}

\item{title,}{y.label Title and Y axis label of the plot.}
}
\description{
For general time series plotting.
}
\examples{
\dontrun{
trades <- lob.data$trades
with(trades, plot.time.series(timestamp, price))

timestamp <- seq(as.POSIXct("2015-01-01 00:00:00.000", tz="UTC"),
                 as.POSIXct("2015-01-01 00:59:00.000", tz="UTC"), by=60)
series <- rep(1:10, 6)
p <- plot.time.series(timestamp, series)
p
}
}

