% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.trades}
\alias{plot.trades}
\title{Plot trades.}
\usage{
\method{plot}{trades}(trades, start.time = min(trades$timestamp),
  end.time = max(trades$timestamp))
}
\arguments{
\item{trades}{Trades data.frame}

\item{start.time}{Plot from}

\item{end.time}{Plot to}
}
\description{
A convenience function for plotting the trades data.frame in a nice way.
}
\examples{
\dontrun{
with(lob.data, plot.trades(trades))
}
}
\author{
phil
}

