% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plotTrades}
\alias{plotTrades}
\title{plotTrades.}
\usage{
plotTrades(trades, start.time = min(trades$timestamp),
  end.time = max(trades$timestamp))
}
\arguments{
\item{trades}{\code{\link{trades}} data.}

\item{start.time}{Plot from.}

\item{end.time}{Plot to.}
}
\description{
A convenience function for plotting the trades data.frame in a nice way.
}
\examples{
with(lob.data, plotTrades(trades))
}
\author{
phil
}

