% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{processData}
\alias{processData}
\title{Import CSV file containing limit order events.}
\usage{
processData(csv.file)
}
\arguments{
\item{csv.file}{Location of CSV file to import}
}
\value{
A list of order book events, trades, order book depth and summary
}
\description{
Imports and performs pre-processing of limit order data contained in a CSV.
}
\examples{
\dontrun{

csv.file <- system.file("extdata", "orders.csv", package="microstructure2")
lob.data <- processData(csv.file)

}
}
\author{
phil
}

