% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\name{saveData}
\alias{saveData}
\title{Save processed data.}
\usage{
saveData(lob.data, bin.file)
}
\arguments{
\item{lob.data}{Limit order, trade and depth data structure.}

\item{bin.file}{File location.}
}
\description{
Saves processed data to file.
}
\author{
phil
}

