% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/needleman-wunsch.R
\name{sMatrix}
\alias{sMatrix}
\title{Similarity matrix.}
\usage{
sMatrix(a, b, filter = function(f1, f2) ifelse(f1 == f2, 1, -1))
}
\arguments{
\item{a}{Vector a.}

\item{b}{Vector b.}

\item{filter}{A similarity function. Default: 1 if equal, -1 otherwise.}
}
\value{
A similarity matrix.
}
\description{
Construct a similarity matrix between 2 vectors.
}
\author{
phil
}
\keyword{internal}

