% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{loadData}
\alias{loadData}
\title{Load pre-processed data.}
\usage{
loadData(bin.file, ...)
}
\arguments{
\item{bin.file}{File location.}

\item{...}{\code{\link{readRDS}}.}
}
\value{
Limit order, trade and depth data structure \code{\link{lob.data}}.
}
\description{
Loads previously saved pre-processed data.
}
\details{
Convenience function.
}
\examples{
\dontrun{

lob.data <- loadData(bin.file="/tmp/lob.data.rds")
}
}
\author{
phil
}

