% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob-analytics.R
\docType{data}
\name{lob.data}
\alias{lob.data}
\title{Example limit order book data.}
\format{A list containing 4 data frames as returned by
        \code{\link{processData}}}
\source{
\url{https://www.bitstamp.net/websocket}
}
\usage{
data(lob.data)
}
\description{
50,393 limit order events. 482 trades.
}
\details{
5 hours of limit order book event data obtained from the Bitstamp (bitcoin)
exchange on 2015-05-01 (midnight until 5am). The data has been preprocessed
with the \code{\link{processData}} function.
}
\author{
phil
}
\references{
\url{https://github.com/phil8192/ticker}
}
\seealso{
\code{\link{events}},
         \code{\link{trades}},
         \code{\link{depth}},
         \code{\link{depth.summary}}
}
\keyword{datasets}

