% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/microstructure2.R
\docType{package}
\name{microstructure2}
\alias{microstructure2}
\alias{microstructure2-package}
\title{microstructure2.}
\description{
package description.
}
\details{
package details.
}
\author{
phil nocturne \email{phil@parasec.net}
Maintainer: Hadley Wickham \email{h.wickham@gmail.com}
}
\references{
\url{http://parasec.net/transmission/order-book-visualisation}
}

