% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot.trades}
\alias{plot.trades}
\title{Plot trades.}
\usage{
\method{plot}{trades}(trades, start.time = min(trades$timestamp),
  end.time = max(trades$timestamp))
}
\arguments{
\item{start.time,}{end.time, Inclusive start and end time of plot.}

\item{timestamp}{POSIXct timestamps.}
}
\description{
Plots executed prices in trades data.frame, where data.frame contains
timestamp and price columns.
}
\examples{
\\donotrun{
x <- load.data("order.book.data.Rda")
p <- plot.trades(x$trades)
p
}
}

