% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/depth.R
\name{priceLevelVolume}
\alias{priceLevelVolume}
\title{Calculate priceLevelVolume (depth).}
\usage{
priceLevelVolume(events)
}
\arguments{
\item{events}{Limit order events.}
}
\value{
Time series of liquidity for each price level in the order book.
}
\description{
Given a data.frame of limit order events, this function will calculate
the cumulative volume for each price level over time. Changes in volume
at a price level can occur when an a new order is added to the queue,
updated (partial fill) or deleted (execution or removal). The resulting
time series is of the form:
\preformatted{
    [timestamp, price, volume, side]
where timestamp = (local) time at which liquidity changed
          price = price level at which liquidity changed
         volume = amount of volume remaining at this price level
           side = the (current) side of the price level in the order book.
}
}
\author{
phil
}

